+++ src/config.h
@@ -48,7 +48,8 @@ typedef struct {
 
 typedef struct {
   // Looks
-  colour fg_colour, bold_colour, bg_colour, cursor_colour, underl_colour;
+  colour fg_colour, bold_colour, bg_colour, cursor_colour;
+  colour underl_colour, hover_colour;
   bool underl_manual;
   colour sel_fg_colour, sel_bg_colour;
   colour search_fg_colour, search_bg_colour, search_current_colour;